refactor(modules/users): move profile endpoint to `UsersProfileController`
