+++ src/modules/users/users.controller.ts
@@ -3,12 +3,10 @@ import {
   Get,
   HttpException,
   HttpStatus,
-  Inject,
   NotFoundException,
   Param,
   ParseUUIDPipe,
   Query,
-  forwardRef,
 } from '@nestjs/common'
 import {
   ApiBadRequestResponse,
@@ -20,7 +18,6 @@ import {
   ApiQuery,
   ApiTags,
 } from '@nestjs/swagger'
-import { firstValueFrom } from 'rxjs'
 
 import { UsersRepository } from './users.repository'
 
@@ -30,13 +27,6 @@ import {
   ONE_IS_INVALID,
   ONE_SUCCESFULLY_FOUND,
 } from '@common/constants'
-import { AuthService } from '@modules/auth'
-import {
-  LastItemQuery,
-  StatisticsService,
-  TopItemQuery,
-} from '@modules/statistics'
-import { ApiItemQuery } from '@modules/statistics/decorators'
 
 export const USER = 'user'
 export const USERS = 'users'
@@ -44,12 +34,7 @@ export const USERS = 'users'
 @Controller(USERS)
 @ApiTags(USERS)
 export class UsersController {
-  constructor(
-    private readonly usersRepository: UsersRepository,
-    @Inject(forwardRef(() => AuthService))
-    private readonly authService: AuthService,
-    private readonly statisticsService: StatisticsService
-  ) {}
+  constructor(private readonly usersRepository: UsersRepository) {}
 
   @Get()
   @ApiOperation({
@@ -73,7 +58,7 @@ export class UsersController {
       return foundUser
     }
 
-    return await this.usersRepository.find()
+    return this.usersRepository.find()
   }
 
   @Get(':id')
@@ -96,141 +81,5 @@ export class UsersController {
     if (!foundUser) throw new NotFoundException(NOT_BEEN_FOUND(USER))
 
     return foundUser
-  }
-
-  @Get(':id/last-tracks')
-  @ApiOperation({
-    summary: "Getting user's last tracks.",
-  })
-  @ApiParam({ name: 'id' })
-  @ApiItemQuery({ withCursors: true })
-  @ApiOkResponse({
-    description: ONE_SUCCESFULLY_FOUND(USER),
-  })
-  @ApiNotFoundResponse({
-    description: NOT_BEEN_FOUND(USER),
-  })
-  @ApiBadRequestResponse({
-    description: ONE_IS_INVALID('uuid'),
-  })
-  async getLastTracks(
-    @Param('id', ParseUUIDPipe) id: string,
-    @Query() { limit, before, after }: LastItemQuery
-  ) {
-    const foundUser = await this.usersRepository.findOneBy({ id })
-
-    if (!foundUser) throw new NotFoundException(NOT_BEEN_FOUND(USER))
-
-    const { accessToken } = await firstValueFrom(
-      this.authService.token({
-        refreshToken: foundUser.refreshToken,
-      })
-    )
-
-    return firstValueFrom(
-      this.statisticsService.lastTracks(accessToken, limit, before, after)
-    )
-  }
-
-  @Get(':id/top-tracks')
-  @ApiOperation({
-    summary: "Getting user's top tracks.",
-  })
-  @ApiParam({ name: 'id' })
-  @ApiItemQuery({ withOffset: true })
-  @ApiOkResponse({
-    description: ONE_SUCCESFULLY_FOUND(USER),
-  })
-  @ApiNotFoundResponse({
-    description: NOT_BEEN_FOUND(USER),
-  })
-  @ApiBadRequestResponse({
-    description: ONE_IS_INVALID('uuid'),
-  })
-  async getTopTracks(
-    @Param('id', ParseUUIDPipe) id: string,
-    @Query() { limit, timeRange, offset }: TopItemQuery
-  ) {
-    const foundUser = await this.usersRepository.findOneBy({ id })
-
-    if (!foundUser) throw new NotFoundException(NOT_BEEN_FOUND(USER))
-
-    const { accessToken } = await firstValueFrom(
-      this.authService.token({
-        refreshToken: foundUser.refreshToken,
-      })
-    )
-
-    return firstValueFrom(
-      this.statisticsService.topTracks(accessToken, limit, timeRange, offset)
-    )
-  }
-
-  @Get(':id/top-genres')
-  @ApiOperation({
-    summary: "Getting user's top genres.",
-  })
-  @ApiParam({ name: 'id' })
-  @ApiItemQuery()
-  @ApiOkResponse({
-    description: ONE_SUCCESFULLY_FOUND(USER),
-  })
-  @ApiNotFoundResponse({
-    description: NOT_BEEN_FOUND(USER),
-  })
-  @ApiBadRequestResponse({
-    description: ONE_IS_INVALID('uuid'),
-  })
-  async getTopGenres(
-    @Param('id', ParseUUIDPipe) id: string,
-    @Query() { limit, timeRange, offset }: TopItemQuery
-  ) {
-    const foundUser = await this.usersRepository.findOneBy({ id })
-
-    if (!foundUser) throw new NotFoundException(NOT_BEEN_FOUND(USER))
-
-    const { accessToken } = await firstValueFrom(
-      this.authService.token({
-        refreshToken: foundUser.refreshToken,
-      })
-    )
-
-    return firstValueFrom(
-      this.statisticsService.topGenres(accessToken, limit, timeRange, offset)
-    )
-  }
-
-  @Get(':id/top-artists')
-  @ApiOperation({
-    summary: "Getting user's top artists.",
-  })
-  @ApiParam({ name: 'id' })
-  @ApiItemQuery({ withOffset: true })
-  @ApiOkResponse({
-    description: ONE_SUCCESFULLY_FOUND(USER),
-  })
-  @ApiNotFoundResponse({
-    description: NOT_BEEN_FOUND(USER),
-  })
-  @ApiBadRequestResponse({
-    description: ONE_IS_INVALID('uuid'),
-  })
-  async getTopArtists(
-    @Param('id', ParseUUIDPipe) id: string,
-    @Query() { limit, timeRange, offset }: TopItemQuery
-  ) {
-    const foundUser = await this.usersRepository.findOneBy({ id })
-
-    if (!foundUser) throw new NotFoundException(NOT_BEEN_FOUND(USER))
-
-    const { accessToken } = await firstValueFrom(
-      this.authService.token({
-        refreshToken: foundUser.refreshToken,
-      })
-    )
-
-    return firstValueFrom(
-      this.statisticsService.topArtists(accessToken, limit, timeRange, offset)
-    )
   }
 }
