+++ src/modules/users/users.module.ts
@@ -4,6 +4,7 @@ import { TypeOrmModule } from '@nestjs/typeorm'
 import { User } from './user.entity'
 import { UsersRepository } from './users.repository'
 import { UsersController } from './users.controller'
+import { UsersProfileController } from './users-profile.controller'
 
 import { AuthModule } from '@modules/auth'
 import { StatisticsModule } from '@modules/statistics'
@@ -15,7 +16,7 @@ import { StatisticsModule } from '@modules/statistics'
     StatisticsModule,
   ],
   providers: [UsersRepository],
-  controllers: [UsersController],
+  controllers: [UsersController, UsersProfileController],
   exports: [UsersRepository],
 })
 export class UsersModule {}
